Move jdbc interface into a separate module
