+++ jdbc-interface/build.gradle
@@ -57,40 +57,28 @@ targetCompatibility = 1.8
 
 dependencies {
     compile project(":core")
-    compile project(":csv-adapter")
-    compile project(":hsqldb-adapter")
-    compile project(":catalog")
-    compile project(":webui")
-    compile project(":jdbc-interface")
 
     ////// Logging
     compile group: "org.apache.logging.log4j", name: "log4j-slf4j-impl", version: "2.11.2" // License: Apache 2.0
     compile group: "org.apache.logging.log4j", name: "log4j-core", version: "2.11.2"   // License: Apache 2.0
     compile group: "org.apache.logging.log4j", name: "log4j-api", version: "2.11.2"
 
-    ////// CLI PARSING
-    compile group: "com.github.rvesse", name: "airline", version: "2.7.0"
-
     ////// APACHE COMMONS LANG
     //compile group: "org.apache.commons", name: "commons-lang3", version: "3.7"
 
     ////// BYTE UNITS
-    //compile group: "com.jakewharton.byteunits", name: "byteunits", version: "0.9.1" // License: Apache 2.0
-    
-    // HSQLDB
-    //compile group: "org.hsqldb", name: "hsqldb", version: "2.5.0"
+    compile group: "com.jakewharton.byteunits", name: "byteunits", version: "0.9.1" // License: Apache 2.0
+
 
     ////// APACHE CALCITE AVATICA
     //compile group: "org.apache.calcite.avatica", name: "avatica-server", version: "1.16.0-POLYPHENYDB"
-    //compile name: "avatica-server-1.16.0-POLYPHENYDB"
+    compile name: "avatica-server-1.16.0-POLYPHENYDB"
     
 
     // --- Test Compile ---
     testCompile project(path: ":core", configuration: "testClasses")
 
     testCompile group: "junit", name: "junit", version:"4.12"
-    
-    testCompile group: "ch.unibas.dmi.dbis.polyphenydb", name: "polyphenydb-jdbc-driver", version:"1.0-SNAPSHOT"
 }
 
 
@@ -117,7 +105,6 @@ sourceSets {
         output.resourcesDir = file(project.buildDir.absolutePath + "/test-classes")
     }
 }
-
 
 
 /**
