+++ jdbc-interface/src/main/java/ch/unibas/dmi/dbis/polyphenydb/jdbc/PolyphenyDbStatementHandle.java
@@ -85,6 +85,4 @@ public class PolyphenyDbStatementHandle {
         this.signature = null;
     }
 
-
-
 }
