+++ settings.gradle
@@ -11,6 +11,7 @@ include 'mongodb-adapter'
 include 'hsqldb-adapter'
 include 'pig-adapter'
 include 'pig-parser'
+include 'jdbc-interface'
 include 'webui'
 include 'dbms'
 include 'catalog'
